refactor(chunks): split signature-aware embedding reads into embeddings_query
Embeddings.rs had grown to 541 lines, over the repo's 500-line ceiling (as
flagged in review). Move the read-side, signature-aware query functions —
get_chunk_embedding_for_signature, get_chunk_embedding,
get_chunk_embeddings_for_signature_batch, get_chunk_embeddings_batch and
has_uncovered_reembed_work (plus the embedding_from_blob decoder and the
MAX_EMBEDDING_BATCH cap) — into a focused embeddings_query sibling.

embeddings.rs keeps the writers, upserts, and re-embed tombstones; both remain
under 500 lines and the public chunks re-export surface is unchanged, so no
caller or test file needs edits.
