+++ src/memory/chunks/mod.rs
@@ -12,9 +12,11 @@
 //! - `semantic` — heading- and paragraph-aware chunker used to split large
 //!   documents into LLM-context-sized pieces while preserving heading context.
 //!   Exported as [`chunk_semantic`].
-//! - `store` / `connection` / `migrations` / `raw_refs` / `embeddings` —
-//!   the SQLite-backed chunk store (the `mem_tree_chunks` table plus its
-//!   per-model embedding sidecars and source ingest gates).
+//! - `store` / `connection` / `migrations` / `raw_refs` / `embeddings` /
+//!   `embeddings_query` — the SQLite-backed chunk store (the `mem_tree_chunks`
+//!   table plus its per-model embedding sidecars and source ingest gates).
+//!   `embeddings` owns the writes and re-embed tombstones; `embeddings_query`
+//!   owns the signature-aware reads and the re-embed coverage probe.
 //!
 //! ## Differences from OpenHuman
 //!
@@ -41,6 +43,8 @@ mod recovery;
 
 #[path = "embeddings.rs"]
 mod embeddings;
+#[path = "embeddings_query.rs"]
+mod embeddings_query;
 #[path = "migrations.rs"]
 mod migrations;
 #[path = "produce.rs"]
@@ -90,12 +94,14 @@ pub use types::{
 pub use connection::{shared_connection, with_connection};
 pub use embeddings::{
     clear_chunk_reembed_skipped, clear_reembed_skipped_for_signature,
-    clear_summary_reembed_skipped, embedding_to_blob, get_chunk_embedding,
-    get_chunk_embedding_for_signature, get_chunk_embeddings_batch,
+    clear_summary_reembed_skipped, embedding_to_blob, mark_chunk_reembed_skipped,
+    mark_summary_reembed_skipped, set_chunk_embedding, set_chunk_embedding_for_signature,
+    set_chunk_embedding_for_signature_tx, set_summary_embedding_for_signature_tx,
+    tree_active_signature, REEMBED_SKIP_KEY_MAX_LEN,
+};
+pub use embeddings_query::{
+    get_chunk_embedding, get_chunk_embedding_for_signature, get_chunk_embeddings_batch,
     get_chunk_embeddings_for_signature_batch, has_uncovered_reembed_work,
-    mark_chunk_reembed_skipped, mark_summary_reembed_skipped, set_chunk_embedding,
-    set_chunk_embedding_for_signature, set_chunk_embedding_for_signature_tx,
-    set_summary_embedding_for_signature_tx, tree_active_signature, REEMBED_SKIP_KEY_MAX_LEN,
 };
 pub use raw_refs::{
     get_chunk_content_path, get_chunk_content_pointers, get_chunk_raw_refs,
